Remove `get_ledger_index`
This commit removes`get_ledger_index()` from ledger traits because `sfLedgerIndex` is not actually a field on a ledger object (it's a synthetic field that maps to the `index` field, which is the unique ID of an object in the ledger's state tree).
See https://github.com/XRPLF/rippled/issues/3649 for more context.